+++ waldreg/stable-service-layer/character/src/test/java/org/waldreg/character/permission/core/PermissionUnitTest.java
@@ -1,4 +1,4 @@
-package org.waldreg.character.core;
+package org.waldreg.character.permission.core;
 
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
