+++ waldreg/stable-service-layer/character/src/test/java/org/waldreg/character/permission/management/PermissionUnitManagerTest.java
@@ -1,4 +1,4 @@
-package org.waldreg.character.management;
+package org.waldreg.character.permission.management;
 
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.waldreg.character.core.PermissionUnit;
+import org.waldreg.character.permission.core.PermissionUnit;
 import org.waldreg.character.exception.DuplicatedPermissionNameException;
 import org.waldreg.character.exception.UnknownPermissionException;
